refactor: 규진 의견 반영해서 모달 animation 조절
