+++ src/components/common/modals/primaryModal/PrimaryModal.tsx
@@ -101,7 +101,7 @@ export default PrimaryModal;
 const StyledReactModal = styled(ReactModal)`
   @keyframes slide {
     from {
-      transform: translateY(0);
+      transform: translateY(-10%);
     }
     to {
       transform: translateY(-50%);
