+++ src/components/common/modals/receiptModal/ReceiptModal.tsx
@@ -154,7 +154,7 @@ export default ReceiptModal;
 const StyledReactModal = styled(ReactModal)`
   @keyframes slide {
     from {
-      transform: translateY(0);
+      transform: translateY(-10%);
     }
     to {
       transform: translateY(-50%);
