+++ src/components/common/modals/secondaryModal/SecondaryModal.tsx
@@ -100,7 +100,7 @@ export default SecondaryModal;
 const StyledReactModal = styled(ReactModal)`
   @keyframes slide {
     from {
-      transform: translateY(0);
+      transform: translateY(-10%);
     }
     to {
       transform: translateY(-50%);
