+++ src/components/home/summaries/KidSummary.tsx
